게시판 기능 구현 중

+++ src/main/java/com/community/communityproject/dto/board/BoardListResponseDTO.java
@@ -18,11 +18,11 @@ public class BoardListResponseDTO {
 
     private String content;
 
-    private Long hits;
+    private int hits;
 
-    private Long reviewCnt;
+    private int reviewCnt;
 
-    private Long likeCnt;
+    private int likeCnt;
 
     private String category;
